Fixes the overlooked init change. Split out messageHandlers.
