+++ lib/handlers/init.js
@@ -60,7 +60,11 @@ var registerNewUser = function registerNewUser(gcmToken, flowdockToken, cb) {
         name: "Welcome to Notiflow!",
       };
 
-      pushNotification(user, "Your " + flows.length + " flows will be monitored.", author, fakeFlow);
+      pushNotification(user, {
+        author: author,
+        flow: fakeFlow,
+        content: "Your " + flows.length + " flows will be monitored."
+      });
       cb(null, user);
     },
   ], cb);
