+++ lib/helpers/listener.js
@@ -5,29 +5,9 @@ var async = require('async');
 var request = require('supertest');
 var rarity = require("rarity");
 
+var messageHandlers = require('./handlers');
 var pushNotification = require('./pusher');
 var toHash = require('./to-hash');
-
-
-
-function handleMessage(message, data) {
-  data.content = message.content.text || message.content;
-}
-
-function handleFile(message, data) {
-  var content = message.content && message.content.file_name;
-  data.content = "[file: " + (content || "attachment") + "]";
-  data.special = true;
-}
-
-/**
- * Message content handlers for notification display
- */
-var messageHandlers = {
-  message:  handleMessage,
-  comment:  handleMessage,
-  file:     handleFile
-};
 
 /*
  * Extract flow data, whether it's a 1 to 1 or a flow
